requirements.txt
pandas
requests
openpyxl
scikit-learn
matplotlib
seaborn
numpy
azure-functions
tensorflow
keras
google-cloud-firestore
google-auth
pyodbc